XMLHttpRequest/send-sync-response-event-order.htm

<!DOCTYPE html>
<html>
<head>
    <link rel="help" href="http://dvcs.w3.org/hg/xhr/raw-file/tip/Overview.html#the-send()-method" />
    <script src="/resources/testharness.js"></script>
    <title>XMLHttpRequest L2: The send() method: event order when synchronous flag is set</title>
</head>

<body>
    <div id="log"></div>

    <script type="text/javascript">
        test(function () {
            var xhr = new XMLHttpRequest();
            var expect = [4, "load", "loadend"];
            var actual = [];

            xhr.onreadystatechange = function() 
            {
                if (xhr.readyState == 4) 
                {
                    actual.push(xhr.readyState);
                }
            };

            xhr.onloadstart        = function(e){ actual.push(e.type); };
            xhr.onload             = function(e){ actual.push(e.type); };
            xhr.onloadend          = function(e){ actual.push(e.type); };

            xhr.upload.onload      = function(e){ actual.push("upload." + e.type); };
            xhr.upload.onloadstart = function(e){ actual.push("upload." + e.type); };
            xhr.upload.onloadend   = function(e){ actual.push("upload." + e.type);};

            xhr.open("POST", "./support/echo.php", false);
            xhr.send("Test Message");

            assert_equals(xhr.response, "Test Message");
            assert_array_equals(actual, expect);
        });
    </script>
</body>
</html>
